refactor: change function name
