+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from "./main/static/api/applyStaticStyles";
+export * from "./main/static/api/useStyle";
 export * from "./main/static/api/requestStyles";
