+++ src/main/client/index.ts
@@ -1,3 +1,3 @@
 export * from "./api/changeScreenMode";
 export * from "./api/saveScreenMode";
-export * from "./api/applyClientStyles";
+export * from "./api/useClientStyle";
