+++ src/main/static/api/useStyle.ts
